* Add last calc fields to zt_metric.

+++ db/update18.10.sql
@@ -1,5 +1,5 @@
 ALTER table `zt_metric` ADD `dateType` varchar(50) NOT NULL DEFAULT '';
-ALTER table `zt_metric` ADD `lastCalcRows` int NOT NULL DEFAULT 0 AFTER `order`;
+ALTER table `zt_metric` ADD `lastCalcRows` mediumint NOT NULL DEFAULT 0 AFTER `order`;
 ALTER table `zt_metric` ADD `lastCalcTime` datetime DEFAULT NULL AFTER `lastCalcRows`;
 
 INSERT INTO `zt_metric`(`purpose`, `scope`, `object`, `stage`, `type`, `name`, `code`, `unit`, `collector`, `desc`, `definition`, `when`, `createdBy`, `createdDate`, `builtin`,  `deleted`, `dateType`) VALUES
